requirements.txt
llama-index-core>=0.10.0
llama-index-llms-llama-cpp>=0.1.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-vector-stores-chroma>=0.1.0
chromadb>=0.4.13
feedparser>=6.0.0
chardet>=5.0.0
fastapi>=0.104.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0
websockets>=11.0.0
psutil>=5.9.0
